compute at least inside cursorToX with floats
BUG: 499669

+++ src/render/katerenderer.h
@@ -304,9 +304,9 @@ public:
     /**
      * Returns the x position of cursor \p col on the line \p range.
      */
-    int cursorToX(const KateTextLayout &range, int col, bool returnPastLine = false) const;
+    qreal cursorToX(const KateTextLayout &range, int col, bool returnPastLine = false) const;
     /// \overload
-    int cursorToX(const KateTextLayout &range, const KTextEditor::Cursor pos, bool returnPastLine = false) const;
+    qreal cursorToX(const KateTextLayout &range, const KTextEditor::Cursor pos, bool returnPastLine = false) const;
 
     /**
      * Returns the real cursor which is occupied by the specified x value, or that closest to it.
